+++ src/App.js
@@ -2,6 +2,7 @@ import Home from "./routes/home/home.component";
 import Navigation from "./routes/navigation/navigation.component";
 import Projects from "./routes/projects/projects.component";
 import { Routes, Route } from "react-router-dom";
+import MyStory from "./routes/home/my-story/my-story.component";
 import "./App.scss";
 
 function Contact() {
@@ -12,18 +13,6 @@ function SiteInfo() {
   return <h1>This site was made with React JS </h1>;
 }
 
-function MyStory() {
-  const imageUrl =
-    "https://upload.wikimedia.org/wikipedia/commons/thumb/8/8d/President_Barack_Obama.jpg/1200px-President_Barack_Obama.jpg";
-  return (
-    <div>
-      <img src={imageUrl} className="profile-pic" />
-      <div className="profile-text">
-        <h1>Hello There!</h1>
-      </div>
-    </div>
-  );
-}
 
 function App() {
   return (
@@ -31,13 +20,7 @@ function App() {
       {/* Render the element if the path matches*/}
       <Route path="/" element={<Navigation />}>
         <Route index element={<Projects />} />
-        <Route path="contact" element={<Contact></Contact>} />
-        <Route path="info" element={<SiteInfo></SiteInfo>} />
-        <Route path="education" element={Contact} />
         <Route path="my-story" element={<MyStory></MyStory>} />
-        <Route path="work-experience" element={Contact} />
-        <Route path="technical-skills" element={Contact} />
-        <Route path="projects" element={<Projects></Projects>} />
       </Route>
     </Routes>
   );
