--- src/App.scss
@@ -1,12 +1,0 @@
-.profile-pic {
-  width: 400px;
-  float: left;
-  padding-top: 15px;
-  vertical-align:bottom;
-}
-
-.profile-text {
-  padding-top: 5px;
-  padding-left: 400px;
-  
-}
