+++ src/routes/home/my-story/my-story.styles.scss
@@ -1,0 +1,12 @@
+.profile-pic {
+  width: 400px;
+  float: left;
+  padding-top: 15px;
+  vertical-align:bottom;
+}
+
+.profile-text {
+  padding-top: 5px;
+  padding-left: 400px;
+  
+}
